fix flake (hopefully)

+++ src/test/common.go
@@ -8,6 +8,7 @@ import (
 	"bufio"
 	"context"
 	"fmt"
+	"math/rand"
 	"os"
 	"path/filepath"
 	"regexp"
@@ -72,15 +73,12 @@ func (e2e *ZarfE2ETest) Zarf(args ...string) (string, string, error) {
 		defer os.RemoveAll(tmpdir)
 		args = append(args, "--tmpdir", tmpdir)
 	}
-	if !slices.Contains(args, "--zarf-cache") && !slices.Contains(args, "tools") {
-		if os.Getenv("CI") == "true" {
-			// We make the cache dir relative to the working directory to make it work on the Windows Runners
-			// - they use two drives which filepath.Rel cannot cope with.
-			relCacheDir, _ := filepath.Abs(".cache-location")
-			args = append(args, "--zarf-cache", relCacheDir)
-			defer os.RemoveAll(relCacheDir)
-		}
-
+	if !slices.Contains(args, "--zarf-cache") && !slices.Contains(args, "tools") && os.Getenv("CI") == "true" {
+		// We make the cache dir relative to the working directory to make it work on the Windows Runners
+		// - they use two drives which filepath.Rel cannot cope with.
+		relCacheDir, _ := filepath.Abs(fmt.Sprintf(".%d-cache-location", rand.Intn(10000000)))
+		args = append(args, "--zarf-cache", relCacheDir)
+		defer os.RemoveAll(relCacheDir)
 	}
 	return exec.CmdWithContext(context.TODO(), exec.PrintCfg(), e2e.ZarfBinPath, args...)
 }
